Auto-Hatchery: fix infinite loop when "Favorite" category is empty
When reaching basicHatchery, if the filtered list is empty, a random
Pokémon is added, but only if Pokémons are not filtered using the
specific "Favorite" category. In that "Favorite" case, the next
iteration of the loop beginning without any new Pokémon in the queue,
and then starts the infinite loop (which freezes the game on my Opera GX
browser).

To fix it:
 * Ensure a return value in basicHatchery: true: if a pokemon was added
   to the queue, otherwise false
 * Check return value, and continue if a Pokémon was added
 * Ensure breaking out at the end of the loop.

For consistency, replace returns with continue, or with break in case of
strict mode (Pokérus only at this time).

Signed-off-by: Motsah <motsah.gaming@gmail.com>